Solution for problem #12
This is a slightly optimized version, my original took almost 20 minutes
to run and this version takes less than 3 seconds.

Technically there is a bug in the code, but that bug will not manifest
itself for the problem's purpose.

The bug is that every divisor found found is counted twice, this was
done to save having to check known divisors and lets me stop when I've
brute forced my way through the sqrt of number we are checking the
divisor count for. However, for the number 1, the only divisor is 1 and
so there will be an off by one error. This case is trivial and so I have
left the bug in the code.

Signed-off-by: Mark Rada <marada@uwaterloo.ca>