Solution for problem #15
I found a pattern in the number of paths that exist
for square grids. This pattern let me implement a very
fast solution to find the number of paths of a square
grid of arbitrary size.

However, I get the feeling there is an even easier way
to solve this problem using combinatorics, but I was too
lazy to try and find it, my brain took the optimal way out
and went with something it could figure out tonight that
wasn't a brute force algorithm.

Signed-off-by: Mark Rada <marada@uwaterloo.ca>